+++ yatr/task.py
@@ -1,9 +1,82 @@
+import re
 import sys
+from itertools import product
+
 from syn.base import Base, Attr
 from syn.type import List, This
 from syn.five import STR
 
-from .base import ValidationError, command
+from .base import ValidationError, command, get_delete
+
+#-------------------------------------------------------------------------------
+
+FOREX = re.compile('([a-zA-Z_][a-zA-Z0-9_-]*)\s+in\s+([a-zA-Z_][a-zA-Z0-9_-]*)')
+
+#-------------------------------------------------------------------------------
+# For
+
+
+class For(Base):
+    _attrs = dict(var = Attr((STR, List(STR)), doc='The loop variable(s)'),
+                  in_ = Attr((STR, List(STR)), doc='Name(s) of list macro(s) to loop over'))
+    _opts = dict(init_validate = True, 
+                 args = ('var', 'in_'))
+
+    @classmethod
+    def from_yaml(cls, dct):
+        if isinstance(dct, STR):
+            m = re.match(FOREX, dct)
+            if m:
+                return cls(var=m.groups()[0], in_=m.groups()[1])
+            else:
+                raise ValidationError('Invalid for loop specifier: {}'
+                                      .format(dct))
+
+        kwargs = {}
+        get_delete(dct, kwargs, 'var', None)
+        get_delete(dct, kwargs, 'in', None, 'in_')
+        
+        if dct:
+            raise ValidationError('Invalid for keys: {}'
+                                  .format(','.join(dct.keys())))
+        return cls(**kwargs)
+
+    def resolve_macros(self, env, **kwargs):
+        var = self.var
+        in_ = self.in_
+        if not isinstance(var, list):
+            var = [self.var]
+            in_ = [self.in_]
+
+        outs = []
+        for k, v in enumerate(var):
+            val = env.env[in_[k]]
+            if not isinstance(val, list):
+                raise ValidationError('For loop "in" specifier must be name of '
+                                      'list macro: {}'.format(in_[k]))
+            outs.append(val)
+        return var, outs
+
+    def loop(self, env, **kwargs):
+        var, in_ = self.resolve_macros(env, **kwargs)
+        
+        for tup in product(*in_):
+            yld = env.copy(**kwargs)
+            for name, val in zip(var, tup):
+                yld.env[name] = val
+            yield yld
+
+    def validate(self):
+        super(For, self).validate()
+        
+        if type(self.var) is not type(self.in_):
+            raise ValidationError('"var" and "in" specifiers must be same type')
+
+        if isinstance(self.var, list):
+            if len(self.var) != len(self.in_):
+                raise ValidationError('"var" and "in" lists must be same '
+                                      'length')
+
 
 #-------------------------------------------------------------------------------
 # Command
@@ -55,6 +128,7 @@ class Command(Base):
 class Task(Base):
     _attrs = dict(commands = Attr(List(Command)),
                   condition = Attr(This, optional=True),
+                  loop = Attr(For, optional=True),
                   condition_type = Attr(bool, True))
     _opts = dict(init_validate = True,
                  optional_none = True)
@@ -76,6 +150,9 @@ class Task(Base):
                     for cmd in ret.commands:
                         cmd.context = dct['context']
 
+                if 'for' in dct:
+                    ret.loop = For.from_yaml(dct['for'])
+
                 if 'if' in dct:
                     ret.condition = Task.from_yaml(name + '-if', dct['if'])
                     ret.condition_type = True
@@ -91,9 +168,10 @@ class Task(Base):
 
     def run(self, env, **kwargs):
         codes = []
+        looping = kwargs.get('looping', False)
         exit_on_error = kwargs.get('exit_on_error', True)
 
-        if self.condition:
+        if self.condition and not looping:
             kwargs['preview_pre'] = 'if: ' if self.condition_type else 'ifnot: '
             codes_ = self.condition.run(env, **kwargs)
             code = max(codes_)
@@ -102,6 +180,13 @@ class Task(Base):
                 code is not None):
                 return []
             kwargs['preview_pre'] = '\t'
+
+        if self.loop and not looping:
+            kwargs['looping'] = True
+            for env_ in self.loop.loop(env, **kwargs):
+                codes_ = self.run(env_, **kwargs)
+                codes.extend(codes_)
+            return codes
 
         for cmd in self.commands:
             if cmd.command in env.tasks:
@@ -121,6 +206,6 @@ class Task(Base):
 #-------------------------------------------------------------------------------
 # __all__
 
-__all__ = ('Command', 'Task')
+__all__ = ('For', 'Command', 'Task')
 
 #-------------------------------------------------------------------------------